ansible/roles/presentation/files/requirements.txt
flask==2.3.2
psycopg2-binary==2.9.6
boto3==1.28.3
python-dotenv==1.0.0
gunicorn==20.1.0
flask-cors==4.0.0
geopandas==0.13.0
shapely==2.0.1
pandas==2.0.3
